requirements.txt
tensorflow>=2.13.0
keras>=3.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pillow>=10.0.0
tqdm>=4.65.0
scipy>=1.11.0
python-dotenv>=1.0.0
